% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{package}
\name{brca.shiny-package}
\alias{brca.shiny}
\alias{brca.shiny-package}
\title{brca.shiny: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
