% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{small_plot}
\alias{small_plot}
\title{Create a small Shiny plot output}
\usage{
small_plot(id, title = id_to_title(id), output = plotly::plotlyOutput)
}
\arguments{
\item{id}{the plot ID as a character string}

\item{title}{(optional) the plot title; if missing, inferred from the `id`}
}
\description{
Create a small Shiny plot output
}
